+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 4.0.0
 * Optimized `TADPole` for multiple `k` and `dc` values.
 * Partitional and hierarchical configurations in `compare_clusterings` now take into account the `symmetric` control parameter if provided.
+* The functionality for `pick.clus` in `compare_clusterings` changed depending on the value of `return.objects`.
 * Optimized PAM centroids with `pam.precompute = FALSE` by using sparse matrices from the `Matrix` package.
 * Fixed an error that sometimes caused objects returned by `tsclust` to have duplicated elements in the `args` slot.
 * Fixed DTW symmetry detection for fuzzy clustering.